Move all extensions to extension folders
