+++ Mapsui.UI.Android/Extensions/PointFExtensions.cs
