+++ Mapsui.UI.Android/MapControl.cs
@@ -9,7 +9,6 @@ using Android.Util;
 using Android.Views;
 using Mapsui.Geometries.Utilities;
 using Mapsui.Logging;
-using Mapsui.Utilities;
 using SkiaSharp.Views.Android;
 using Math = System.Math;
 using Point = Mapsui.Geometries.Point;
