+++ Mapsui/Extensions/BoundingBoxExtensions.cs
@@ -1,6 +1,5 @@
 using BruTile;
 
-// ReSharper disable CheckNamespace
 namespace Mapsui.Geometries
 {
     public static class BoundingBoxExtensions
