+++ Mapsui/Extensions/DoubleExtensions.cs
@@ -1,4 +1,3 @@
- // ReSharper disable once CheckNamespace
 namespace System
 {
     public static class DoubleExtensions
