+++ Samples/Mapsui.Samples.Avalonia/Views/MainWindow.axaml.cs
@@ -5,7 +5,7 @@ using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Input;
 using Avalonia.Markup.Xaml;
-using Mapsui.Samples.Avalonia.ExtensionMethods;
+using Mapsui.Samples.Avalonia.Extensions;
 using Mapsui.Samples.Common;
 using Mapsui.Samples.Common.Helpers;
 using Mapsui.Samples.Common.Maps;
