+++ Samples/Mapsui.Samples.Droid/MainActivity.cs
@@ -8,7 +8,7 @@ using Android.Support.V7.App;
 using Android.Views;
 using Mapsui.Providers;
 using Mapsui.Samples.Common;
-using Mapsui.Samples.Common.ExtensionMethods;
+using Mapsui.Samples.Common.Extensions;
 using Mapsui.Samples.Common.Helpers;
 using Mapsui.Samples.Common.Maps;
 using Mapsui.Samples.CustomWidget;
