+++ Samples/Mapsui.Samples.Forms/Mapsui.Samples.Forms.Shared/MainPageLarge.xaml.cs
@@ -1,7 +1,6 @@
 ﻿using Mapsui.Providers;
 using Mapsui.Rendering.Skia;
 using Mapsui.Samples.Common;
-using Mapsui.Samples.Common.ExtensionMethods;
 using Mapsui.Samples.CustomWidget;
 using Mapsui.UI.Forms;
 using Plugin.Geolocator;
@@ -9,6 +8,7 @@ using Plugin.Geolocator.Abstractions;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using Mapsui.Samples.Common.Extensions;
 using Xamarin.Forms;
 using Xamarin.Forms.Xaml;
