+++ Samples/Mapsui.Samples.Uwp/Extensions/FeatureExtensions.cs
@@ -2,11 +2,8 @@
 using System.Collections.Generic;
 using System.Text;
 using Mapsui.Providers;
-#if __WINUI__
-namespace Mapsui.Samples.WinUI
-#else
-namespace Mapsui.Samples.Uwp
-#endif
+
+namespace Mapsui.Samples.Uwp.Extensions
 {
     static class FeatureExtensions
     {
