+++ Samples/Mapsui.Samples.Uwp/MainPage.xaml.cs
@@ -11,6 +11,7 @@ using Mapsui.Samples.Common;
 using Mapsui.Samples.Common.Helpers;
 using Mapsui.Samples.Common.Maps;
 using Mapsui.Samples.CustomWidget;
+using Mapsui.Samples.Uwp.Extensions;
 
 namespace Mapsui.Samples.Uwp
 {
