+++ Samples/Mapsui.Samples.Uwp/Mapsui.Samples.Uwp.csproj
@@ -94,7 +94,7 @@
     <Compile Include="App.xaml.cs">
       <DependentUpon>App.xaml</DependentUpon>
     </Compile>
-    <Compile Include="ExtensionMethods\FeatureExtensions.cs" />
+    <Compile Include="Extensions\FeatureExtensions.cs" />
     <Compile Include="MainPage.xaml.cs">
       <DependentUpon>MainPage.xaml</DependentUpon>
     </Compile>
