+++ Samples/Mapsui.Samples.WinUI/Mapsui.Samples.WinUI/Mapsui.Samples.WinUI.csproj
@@ -10,9 +10,6 @@
     <UseWinUI>true</UseWinUI>
     <DefineConstants>__WINUI__</DefineConstants>
   </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="..\..\Mapsui.Samples.Uwp\ExtensionMethods\FeatureExtensions.cs" Link="ExtensionMethods\FeatureExtensions.cs" />
-  </ItemGroup>
 
   <ItemGroup>
     <PackageReference Include="Microsoft.ProjectReunion" Version="0.8.4" />
@@ -21,10 +18,6 @@
     <PackageReference Include="SkiaSharp" Version="2.80.3" />
     <PackageReference Include="SkiaSharp.Views.WinUI" Version="2.80.3" />
     <Manifest Include="$(ApplicationManifest)" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="ExtensionMethods\" />
   </ItemGroup>
 
   <ItemGroup>
