+++ Samples/Mapsui.Samples.Wpf.Editing/Extensions/IGeometryExtensions.cs
