+++ Samples/Mapsui.Samples.Wpf.Editing/Extensions/InfoEventArgsExtensions.cs
@@ -1,5 +1,6 @@
 ﻿using System.Windows.Navigation;
 using Mapsui.Providers;
+using Mapsui.Samples.Wpf.Editing.Extensions;
 using Mapsui.UI;
 
 namespace Mapsui.Samples.Wpf.Editing
