+++ Samples/Mapsui.Samples.Wpf/Extensions/FeatureExtensions.cs
